bot/requirements.txt
aiomysql==0.2.0
cryptography==41.0.2
python-telegram-bot==20.4